refactor(web): extract dashboard API data types
