+++ web/src/App.tsx
@@ -106,7 +106,7 @@ import { useTheme } from "@/themes";
 import { isDashboardEmbeddedChatEnabled } from "@/lib/dashboard-flags";
 import { latchChatActivation } from "@/lib/chat-activation";
 import { api } from "@/lib/api";
-import type { StatusResponse, UpdateCheckResponse } from "@/lib/api";
+import type { StatusResponse, UpdateCheckResponse } from "@/lib/api-types";
 
 function RouteFallback({ label = "Loading…" }: { label?: string }) {
   return (
