+++ web/src/components/OAuthLoginModal.tsx
@@ -3,7 +3,8 @@ import { ExternalLink, X, Check, Copy } from "lucide-react";
 import { Button } from "@nous-research/ui/ui/components/button";
 import { Spinner } from "@nous-research/ui/ui/components/spinner";
 import { H2 } from "@nous-research/ui/ui/components/typography/h2";
-import { api, type OAuthProvider, type OAuthStartResponse } from "@/lib/api";
+import { api } from "@/lib/api";
+import type { OAuthProvider, OAuthStartResponse } from "@/lib/api-types";
 import { copyTextToClipboard } from "@/lib/clipboard";
 import { Input } from "@nous-research/ui/ui/components/input";
 import { useI18n } from "@/i18n";
