+++ web/src/components/SidebarFooter.tsx
@@ -1,5 +1,5 @@
 import { Typography } from "@nous-research/ui/ui/components/typography/index";
-import type { StatusResponse } from "@/lib/api";
+import type { StatusResponse } from "@/lib/api-types";
 import { cn } from "@/lib/utils";
 import { useI18n } from "@/i18n";
