+++ web/src/pages/AnalyticsPage.tsx
@@ -15,7 +15,7 @@ import type {
   AnalyticsDailyEntry,
   AnalyticsModelEntry,
   AnalyticsSkillEntry,
-} from "@/lib/api";
+} from "@/lib/api-types";
 import { timeAgo } from "@/lib/utils";
 import { Button } from "@nous-research/ui/ui/components/button";
 import { Spinner } from "@nous-research/ui/ui/components/spinner";
